+++ include/nfx/silicon/Mux.h
@@ -1,0 +1,8 @@
+#pragma once
+
+/**
+ * \file Mux.h
+ * \brief Convenience header for all multiplexer packages.
+ */
+
+#include "mux/cd4xxx/CD4051B.h"